+++ xenon-common/src/main/java/com/vmware/xenon/services/common/NodeGroupService.java
@@ -187,6 +187,11 @@ public class NodeGroupService extends StatefulService {
         }
 
         NodeGroupState localState = getState(patch);
+        if (localState == null || localState.nodes == null) {
+            logWarning("Invalid local state");
+            patch.fail(Operation.STATUS_CODE_FAILURE_THRESHOLD);
+            return;
+        }
 
         if (body.config == null && body.nodes.isEmpty()) {
             UpdateQuorumRequest bd = patch.getBody(UpdateQuorumRequest.class);
@@ -313,9 +318,16 @@ public class NodeGroupService extends StatefulService {
             return;
         }
 
+        NodeGroupState localState = getState(post);
+        if (localState == null || localState.nodes == null) {
+            logWarning("invalid local state");
+            post.fail(Operation.STATUS_CODE_BAD_REQUEST);
+            return;
+        }
+
         CheckConvergenceRequest cr = post.getBody(CheckConvergenceRequest.class);
         if (CheckConvergenceRequest.KIND.equals(cr.kind)) {
-            handleCheckConvergencePost(post, cr);
+            handleCheckConvergencePost(post, localState, cr);
             return;
         }
 
@@ -348,14 +360,12 @@ public class NodeGroupService extends StatefulService {
             body.documentSelfLink = UriUtils.buildUriPath(getSelfLink(), body.id);
         }
 
-        NodeGroupState localState = getState(post);
         localState.nodes.put(body.id, body);
-
         post.setBody(localState).complete();
     }
 
-    private void handleCheckConvergencePost(Operation post, CheckConvergenceRequest body) {
-        NodeGroupState localState = getState(post);
+    private void handleCheckConvergencePost(Operation post, NodeGroupState localState,
+            CheckConvergenceRequest body) {
         CheckConvergenceResponse rsp = new CheckConvergenceResponse();
         rsp.isConverged = localState.membershipUpdateTimeMicros == body.membershipUpdateTimeMicros;
         post.setBody(rsp).complete();